Use CI="true" instead of CONTENT_DIR="content".

+++ bin/deploy.sh
@@ -7,8 +7,11 @@ npm run build
 touch public/.nojekyll
 echo 'sijiaoh.com' > public/CNAME
 
-rm -rf sijiaoh.github.io
-git clone -b gh-pages --depth 1 git@github.com:sijiaoh/sijiaoh.github.io.git
+if [ "${CI:-}" != "true" ]; then
+  rm -rf sijiaoh.github.io
+  git clone -b gh-pages --depth 1 git@github.com:sijiaoh/sijiaoh.github.io.git
+fi
+
 cd sijiaoh.github.io
 git config --local user.email "sijiaoh@outlook.com"
 git config --local user.name "sijiaoh"
